mlp et linear qui compile

+++ lib/linear_model.h
@@ -5,11 +5,11 @@
 #ifndef LIB_LINEAR_MODEL_H
 #define LIB_LINEAR_MODEL_H
 
-extern "C" double* create_linear_model(int);
-extern "C" short predict_linear_classification(double *, int, double *);
-extern "C" double predict_linear_model_regression(double *, int, double *);
-extern "C" double* train_linear_classification(double *, double *, double*, int , int );
-extern "C" double get_random_once();
-extern "C" double* train_regression_model(double*, int, double *, int, double *, int);
-extern "C" double* train_rosenblatt_linear_model(double *, int, double *, int, double *, int , float);
+extern "C" float* create_linear_model(int);
+extern "C" short predict_linear_classification(float *, int, float *);
+extern "C" float predict_linear_model_regression(float *, int, float *);
+extern "C" float* train_linear_classification(float *, float *, float*, int , int );
+extern "C" float get_random_once();
+extern "C" float* train_regression_model(float*, int, float *, int, float *, int);
+extern "C" float* train_rosenblatt_linear_model(float *, int, float *, int, float *, int , float);
 #endif //LIB_LINEAR_MODEL_H
